+++ services/sso/src/__tests__/users/list.test.ts
@@ -1,5 +1,5 @@
 import supertest from 'supertest'
-import { describe, expect, it } from 'vitest'
+import { afterEach, describe, expect, it } from 'vitest'
 import qs from 'qs'
 import { userSchema } from '../../schemas'
 import { server, testUserData } from '../globalSetup'
@@ -15,15 +15,25 @@ const userIdData = await client.query(
   'SELECT id FROM "user" WHERE dni IN ($1) ',
   [testUserData.user.dni]
 )
+const deletedUserIdData = await client.query(
+  'SELECT id FROM "user" WHERE dni IN ($1) ',
+  [testUserData.userToDelete.dni]
+)
 const { id } = adminIdData.rows[0]
 const { id: userId } = userIdData.rows[0]
+const { id: deletedUserId } = deletedUserIdData.rows[0]
 const responseSchema = userSchema.pick({ id: true, name: true }).array()
 const stringData = (data: string[]) =>
   qs.stringify(
     { id: data, fields: ['id', 'name'] },
     { indices: false, arrayFormat: 'comma' }
   )
-
+afterEach(async () => {
+  await client.query('UPDATE "user" SET deleted_at = $1 WHERE dni = $2', [
+    null,
+    testUserData.userToDelete.dni,
+  ])
+})
 describe('Testing get users name by Id endpoint', () => {
   it('returns a user successfully with a single valid ID', async () => {
     const response = await supertest(server).get(`${route}?${stringData(id)}`)
@@ -31,6 +41,7 @@ describe('Testing get users name by Id endpoint', () => {
     expect(response.body).toHaveLength(1)
     expect(responseSchema.safeParse(response.body).success).toBeTruthy()
   })
+
   it('returns multiple users successfully with multiple valid IDs', async () => {
     const response = await supertest(server).get(
       `${route}?${stringData([id, userId])}`
@@ -39,6 +50,7 @@ describe('Testing get users name by Id endpoint', () => {
     expect(response.body).toHaveLength(2)
     expect(responseSchema.safeParse(response.body).success).toBeTruthy()
   })
+
   it('returns a single user for duplicate IDs in the request', async () => {
     const response = await supertest(server).get(
       `${route}?${stringData([id, id, id])}`
@@ -47,6 +59,7 @@ describe('Testing get users name by Id endpoint', () => {
     expect(response.body).toHaveLength(1)
     expect(responseSchema.safeParse(response.body).success).toBeTruthy()
   })
+
   it('returns a empty array when the ID is not found', async () => {
     const response = await supertest(server).get(
       `${route}?${stringData(['m9ftb049xe9sit7155al0mpk'])}`
@@ -54,6 +67,7 @@ describe('Testing get users name by Id endpoint', () => {
     expect(response.status).toBe(200)
     expect(response.body).toHaveLength(0)
   })
+
   it('fails with a Zod validation error when ID is an empty string', async () => {
     const response = await supertest(server).get(`${route}?${stringData([''])}`)
     expect(response.status).toBe(400)
@@ -67,5 +81,29 @@ describe('Testing get users name by Id endpoint', () => {
         },
       ],
     })
+  })
+
+  it('returns an empty array if the user is deleted', async () => {
+    await client.query(
+      'UPDATE "user" SET deleted_at = CURRENT_TIMESTAMP WHERE dni = $1',
+      [testUserData.userToDelete.dni]
+    )
+    const response = await supertest(server).get(
+      `${route}?${stringData([deletedUserId])}`
+    )
+    expect(response.status).toBe(200)
+    expect(response.body).toHaveLength(0)
+  })
+
+  it('returns just the not deleted user', async () => {
+    await client.query(
+      'UPDATE "user" SET deleted_at = CURRENT_TIMESTAMP WHERE dni = $1',
+      [testUserData.userToDelete.dni]
+    )
+    const response = await supertest(server).get(
+      `${route}?${stringData([deletedUserId, id])}`
+    )
+    expect(response.status).toBe(200)
+    expect(response.body).toHaveLength(1)
   })
 })