+++ web/usuaris/src/types/index.tsx
@@ -1,3 +1,4 @@
-export { type TUserData } from './types'
+export { type TFilters } from './types'
 export { type TItinerary } from './types'
 export { type TUpdatedUser } from './types'
+export { type TUserData } from './types'